% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_pipelines.R
\docType{data}
\name{AnnPhaseVcf}
\alias{AnnPhaseVcf}
\title{AnnPhaseVcf}
\format{A `cwlStepParam` object.
\describe{
 steps:
 \item{VCFvep}{annotate VCF with VEP}
 \item{VCFcoverage}{add reads counts to VCF}
 \item{VCFexpression}{add expression data to VCF}
 \item{phaseVcf}{combine germline and somatic variant and phase}
}}
\source{
\url{https://pvactools.readthedocs.io/en/latest/pvacseq/input_file_prep.html}
}
\usage{
AnnPhaseVcf
}
\description{
AnnPhaseVcf
}
\keyword{datasets}
