% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_tools.R
\docType{data}
\name{bam_readcount}
\alias{bam_readcount}
\title{bam_readcount}
\format{An object of class \code{cwlParam} of length 1.}
\usage{
bam_readcount
}
\description{
bam_readcount
}
\keyword{datasets}
