% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_tools.R
\docType{data}
\name{bamCoverage}
\alias{bamCoverage}
\title{deeptools bamCoverage}
\format{An object of class \code{cwlParam} of length 1.}
\usage{
bamCoverage
}
\description{
deeptools bamCoverage
}
\keyword{datasets}
