% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_pipelines.R
\docType{data}
\name{BaseRecal}
\alias{BaseRecal}
\title{Base quality recalibration}
\format{A `cwlStepParam` object.
\describe{
 \item{BaseRecalibrator}{Detect systematic errors in base quality scores}
 \item{ApplyBQSR}{Apply base quality score recalibration}
 \item{samtools_index}{samtools index}
 \item{samtools_flagstat}{samtools flagstat}
 \item{samtools_stats}{samtools stats}
}}
\source{
\url{https://software.broadinstitute.org/gatk/best-practices/workflow?id=11165}
}
\usage{
BaseRecal
}
\description{
Base quality recalibration
}
\keyword{datasets}
