% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_tools.R
\docType{data}
\name{BedToIntervalList}
\alias{BedToIntervalList}
\title{Picard BedToIntervalList}
\format{An object of class \code{cwlParam} of length 1.}
\usage{
BedToIntervalList
}
\description{
Picard BedToIntervalList
}
\keyword{datasets}
