% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_tools.R
\docType{data}
\name{bowtie2_build}
\alias{bowtie2_build}
\title{bowtie2-build}
\format{An object of class \code{cwlParam} of length 1.}
\usage{
bowtie2_build
}
\description{
bowtie2-build
}
\keyword{datasets}
