% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_pipelines.R
\docType{data}
\name{bwaAlign}
\alias{bwaAlign}
\title{bwaAlign}
\format{A `cwlStepParam` object.
\describe{
 \item{bwa}{to align fastqs with bwa}
 \item{sam2bam}{samtools view sam to bam format}
 \item{sortBam}{sort Bam file by samtools sort}
 \item{idxBam}{index Bam by samtools index}
}}
\usage{
bwaAlign
}
\description{
bwa alignment
}
\keyword{datasets}
