% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_pipelines.R
\docType{data}
\name{bwaMMRecal}
\alias{bwaMMRecal}
\title{DNASeq alignment, merge, markduplicates and recalibration}
\format{A `cwlStepParam` object.
\describe{
 \item{bwaAlign}{to align fastqs with bwa and sort with samtools}
 \item{mergeBamDup}{to merge BAMs from different flowcells
and then mark duplicates with picard}
 \item{BaseRecal}{Base quality recalibration}
}}
\source{
\url{https://hubentu.github.io/others/Rcwl_DNASeq_Align.html}
}
\usage{
bwaMMRecal
}
\description{
The DNASeq pipeline to run bwa alignment, merge, mark
duplicates and recalibration.
}
\keyword{datasets}
