% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_pipelines.R
\docType{data}
\name{bwaMRecal}
\alias{bwaMRecal}
\title{DNASeq alignment, markduplicates and recalibration}
\format{A `cwlStepParam` object.
\describe{
 \item{bwaAlign}{to align fastqs with bwa and sort with samtools}
 \item{markdup}{to mark duplicates with picard}
 \item{BaseRecal}{Base quality recalibration}
}}
\source{
\url{https://hubentu.github.io/others/Rcwl_DNASeq_Align.html}
}
\usage{
bwaMRecal
}
\description{
The DNASeq pipeline to run bwa alignment, mark duplicates and
recalibration.
}
\keyword{datasets}
