% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_tools.R
\docType{data}
\name{ColSeqArtifact}
\alias{ColSeqArtifact}
\title{gatk CollectSequencingArtifactMetrics}
\format{An object of class \code{cwlParam} of length 1.}
\usage{
ColSeqArtifact
}
\description{
gatk CollectSequencingArtifactMetrics
}
\keyword{datasets}
