% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_tools.R
\docType{data}
\name{CombineVariants}
\alias{CombineVariants}
\title{CombineVariants}
\format{An object of class \code{cwlParam} of length 1.}
\usage{
CombineVariants
}
\description{
CombineVariants
}
\keyword{datasets}
