% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allclasses.R
\docType{class}
\name{cwlHub-class}
\alias{cwlHub-class}
\alias{cwlHub}
\alias{mcols,cwlHub-method}
\alias{show,cwlHub-method}
\alias{[,cwlHub,ANY,ANY,ANY-method}
\alias{title}
\alias{Command}
\alias{Container}
\alias{Type}
\title{cwlHub}
\usage{
cwlHub(BFC)

\S4method{mcols}{cwlHub}(x)

\S4method{show}{cwlHub}(object)

\S4method{[}{cwlHub,ANY,ANY,ANY}(x, value)

title(object)

Command(object)

Container(object)

Type(object)
}
\arguments{
\item{BFC}{A BiocFileCache created for `RcwlRecipes`.}

\item{x}{A `cwlHub` object.}

\item{object}{A `cwlHub` object}

\item{value}{The "BFC" ID to extract the subset.}
}
\value{
cwlHub: a `cwlHub` object with slots of `rid` and `cache` path.

mcols: a `DataFrame` with information from the `BicFileCache` object.

[: a subset of `cwlHub` records.

title: the `Rcwl` recipe names for tools or pipelines.

Command: The name of `Rcwl` wrapped command line tools.

Container: the container name for the `Rcwl` recipe if exist. Otherwise `NA`.

Type: The type of the `Rcwl` recipe, "pipeline" or "tool".
}
\description{
`cwlHub` class, constructor, and methods.
}
\examples{
\dontrun{
tools <- cwlUpdate()
t1 <- tools["BFC178"]
title(t1)
Command(t1)
Container(t1)
Type(t1)
}

}
