% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allclasses.R
\name{title}
\alias{title}
\alias{Command}
\alias{Container}
\alias{Type}
\title{title}
\usage{
title(object)

Command(object)

Container(object)

Type(object)
}
\arguments{
\item{object}{A `cwlHub` object.}
}
\description{
title

Command

Container

Type
}
\examples{
\dontrun{
tools <- cwlUpdate()
title(tools)
}
}
