% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwlInstall.R
\name{cwlInstall}
\alias{cwlInstall}
\title{cwlInstall}
\usage{
cwlInstall(file, rname, bfc = NULL, env = .GlobalEnv)
}
\arguments{
\item{file}{The Rcwl tool or pipeline receipes. The dependent
Rcwl scripts should be included with @include tag.}

\item{rname}{The `rname` to install from the `bfc` object.}

\item{bfc}{The BiocFileCache object for the recipes.}

\item{env}{The enviroment to export.}
}
\description{
To source Rcwl scripts
}
\examples{
\dontrun{
cwlInstall(rname = "tl_bwa")
}
}
