% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwlSearch.R
\name{cwlSearch}
\alias{cwlSearch}
\title{cwlSearch}
\usage{
cwlSearch(keyword, bfc = NULL, ...)
}
\arguments{
\item{keyword}{The keyword to search.}

\item{bfc}{The BiocFileCache object for the recipes.}

\item{...}{More options from `bfcquery`.}
}
\value{
A BiocFileCache tibble.
}
\description{
Function to search Rcwl tools and pipelines.
}
\examples{
\dontrun{
cwlSearch("bwa")
}
}
