% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwlTools.R
\name{cwlTools}
\alias{cwlTools}
\title{cwlTools}
\usage{
cwlTools(cachePath = "Rcwl", ...)
}
\arguments{
\item{cachePath}{The cache path of the BiocFileCache object.}

\item{...}{options from `bfcadd`.}
}
\value{
A BiocFileCache object for existing CWL tools.
}
\description{
To generate a file cache object for CWL tools in the package.
}
\examples{
tools <- cwlTools()
}
