% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwlUpdate.R
\name{cwlUpdate}
\alias{cwlUpdate}
\title{cwlUpdate}
\usage{
cwlUpdate(cachePath = "Rcwl", force = FALSE, branch = "master")
}
\arguments{
\item{cachePath}{The cache path of the BiocFileCache object to
store the Rcwl tools and pipelines recipes.}

\item{force}{Whether to clean existing recipes cache.}

\item{branch}{The branch of github recipes repository. It can be
"master" and "dev". "force = TRUE" is recommended when swithing
branch.}
}
\description{
Function to update the Rcwl recipes from the RcwlRecipes github
repository.
}
\examples{
\dontrun{
tools <- cwlUpdate()
}
}
