% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_tools.R
\docType{data}
\name{DepthOfCoverage}
\alias{DepthOfCoverage}
\title{Picard DepthOfCoverage}
\format{An object of class \code{cwlParam} of length 1.}
\usage{
DepthOfCoverage
}
\description{
Picard DepthOfCoverage
}
\keyword{datasets}
