% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_tools.R
\docType{data}
\name{FilterMutectCalls}
\alias{FilterMutectCalls}
\title{gatk FilterMutectCalls}
\format{An object of class \code{cwlParam} of length 1.}
\usage{
FilterMutectCalls
}
\description{
gatk FilterMutectCalls
}
\keyword{datasets}
