% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_tools.R
\docType{data}
\name{FilterOBias}
\alias{FilterOBias}
\title{gatk FilterByOrientationBias}
\format{An object of class \code{cwlParam} of length 1.}
\usage{
FilterOBias
}
\description{
gatk FilterByOrientationBias
}
\keyword{datasets}
