% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAlign.R
\docType{data}
\name{GAlign}
\alias{GAlign}
\title{GATK alignment pipeline}
\format{A `cwlStepParam` object.
\describe{
 \item{fq2ubam}{To covert fastq to ubam with read group information}
 \item{align}{To run BWA alignment and BAM BaseRecalibration.}
}}
\source{
\url{https://github.com/gatk-workflows/seq-format-conversion}

\url{https://github.com/gatk-workflows/gatk4-data-processing}

\url{https://hubentu.github.io/others/Rcwl_GATK4.html}
}
\usage{
GAlign
}
\description{
Workflows for processing high-throughput sequencing data for
variant discovery with GATK4 and related tools.
}
\examples{
## scripts to build the pipeline
demo("GAlign")
}
\keyword{datasets}
