% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_tools.R
\docType{data}
\name{geneBody_coverage}
\alias{geneBody_coverage}
\title{geneBody_coverage.py}
\format{An object of class \code{cwlParam} of length 1.}
\usage{
geneBody_coverage
}
\description{
geneBody_coverage.py
}
\keyword{datasets}
