% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_tools.R
\docType{data}
\name{genePredToBed}
\alias{genePredToBed}
\title{genePredToBed}
\format{An object of class \code{cwlParam} of length 1.}
\usage{
genePredToBed
}
\description{
genePredToBed
}
\keyword{datasets}
