% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_tools.R
\docType{data}
\name{GenomicsDB}
\alias{GenomicsDB}
\title{gatk GenomicsDBImport}
\format{An object of class \code{cwlParam} of length 1.}
\usage{
GenomicsDB
}
\description{
gatk GenomicsDBImport
}
\keyword{datasets}
