% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_tools.R
\docType{data}
\name{GetPileupSummaries}
\alias{GetPileupSummaries}
\title{gatk GetPileupSummaries}
\format{An object of class \code{cwlParam} of length 1.}
\usage{
GetPileupSummaries
}
\description{
gatk GetPileupSummaries
}
\keyword{datasets}
