% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_pipelines.R
\docType{data}
\name{GPoN}
\alias{GPoN}
\title{GATK4: create a panel of normals}
\format{A `cwlStepParam` object.
\describe{
 \item{GPoN}{The best practice pipeline to create a panel of normals.}
}}
\source{
\url{https://software.broadinstitute.org/gatk/documentation/article?id=24057}
}
\usage{
GPoN
}
\description{
The Panel of Normals Workflow
}
\keyword{datasets}
