% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_tools.R
\docType{data}
\name{gtfToGenePred}
\alias{gtfToGenePred}
\title{gtfToGenePred}
\format{An object of class \code{cwlParam} of length 1.}
\usage{
gtfToGenePred
}
\description{
gtfToGenePred
}
\keyword{datasets}
