% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hapCall.R
\docType{data}
\name{hapCall}
\alias{hapCall}
\title{GATK haplotypecaller pipeline}
\format{A `cwlStepParam` object.
\describe{
 \item{HC}{HaplotypeCaller from GATK4}
}}
\source{
\url{https://github.com/gatk-workflows/gatk4-germline-snps-indels}

\url{https://hubentu.github.io/others/Rcwl_GATK4.html}
}
\usage{
hapCall
}
\description{
The workflow runs HaplotypeCaller from GATK4 in GVCF mode on a
single sample according to the GATK Best Practices (June 2016),
scattered across intervals. The workflow from github,
gatk4-germline-snps-indels (last update: 7/23/2018) was cloned to
this package.
}
\keyword{datasets}
