% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdCall.R
\docType{data}
\name{jdCall}
\alias{jdCall}
\title{GATK joint discovery pipeline}
\format{A `cwlStepParam` object.
\describe{
 \item{JD}{variant joint genotyping}
}}
\source{
\url{https://github.com/gatk-workflows/gatk4-germline-snps-indels}

\url{https://hubentu.github.io/others/Rcwl_GATK4.html}
}
\usage{
jdCall
}
\description{
The joint discovery and VQSR filtering portion of the GATK Best
Practices (June 2016) for germline SNP and Indel discovery in human
whole-genome sequencing (WGS) and exome sequencing data.
}
\examples{
## scripts to build the pipeline
## demo("jdCall")
}
\keyword{datasets}
