% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_tools.R
\docType{data}
\name{kallisto_index}
\alias{kallisto_index}
\title{Kallisto index}
\format{An object of class \code{cwlParam} of length 1.}
\usage{
kallisto_index
}
\description{
Kallisto index
}
\keyword{datasets}
