% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_tools.R
\docType{data}
\name{kallisto_quant}
\alias{kallisto_quant}
\title{Kallisto quant}
\format{An object of class \code{cwlParam} of length 1.}
\usage{
kallisto_quant
}
\description{
Kallisto quant
}
\keyword{datasets}
