% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_pipelines.R
\docType{data}
\name{mantaStrelka}
\alias{mantaStrelka}
\title{strelka somatic caller}
\format{A `cwlStepParam` object.
\describe{
 steps:
 \item{manta}{Call candidate small indels}
 \item{strelka}{somatic calling by strelka2}
}}
\source{
\url{https://github.com/Illumina/strelka}
}
\usage{
mantaStrelka
}
\description{
Strelka2 Somatic caller pipeline.
}
\keyword{datasets}
