% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc3.R
\docType{data}
\name{mc3}
\alias{mc3}
\title{TCGA MC3 pipeline}
\format{A `cwlStepParam` object.
\describe{
 \item{call_variants}{To call somatic variants by multiple pipelines}
 \item{convert}{To convert and merge VCFs to MAF}
}}
\source{
\url{https://github.com/OpenGenomics/mc3}

\url{https://hubentu.github.io/others/Rcwl_MC3.html}
}
\usage{
mc3
}
\description{
The MC3 somatic variant calling pipeline.
}
\examples{
## scripts to build the pipeline
demo("mc3")
}
\keyword{datasets}
