% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allclasses.R
\name{mcols,cwlHub-method}
\alias{mcols,cwlHub-method}
\title{mcols}
\usage{
\S4method{mcols}{cwlHub}(x)
}
\arguments{
\item{x}{A `cwlHub` object}
}
\description{
DataFrame information from the `BicFileCache` object.
}
