% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_pipelines.R
\docType{data}
\name{mergeBamDup}
\alias{mergeBamDup}
\title{mergeBamDup}
\format{A `cwlStepParam` object.
\describe{
 \item{mergeBam}{picard merge Bam files}
 \item{markdup}{picard mark duplicated alignments}
 \item{samtools_index}{index Bam by samtools index}
 \item{samtools_flagstat}{samtools flagstat}
}}
\usage{
mergeBamDup
}
\description{
merge Bam files and mark duplicates
}
\keyword{datasets}
