% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_pipelines.R
\docType{data}
\name{miRDeep2PL}
\alias{miRDeep2PL}
\title{miRDeep2}
\format{A `cwlStepParam` object.
\describe{
 steps:
 \item{miRMapper}{run mapple.pl}
 \item{miRDeep2}{run mirdeep2.pl}
}}
\source{
\url{https://github.com/rajewsky-lab/mirdeep2}
}
\usage{
miRDeep2PL
}
\description{
miRDeep2
}
\keyword{datasets}
