% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiqc.R
\docType{data}
\name{multiqc}
\alias{multiqc}
\title{multiqc}
\format{A `cwlParam` object.
\describe{
 \item{bwaAlign}{to align with bwa and sort with samtools}
 \item{mergeBamDup}{to merge BAMs from different flowcells and then mark duplicates with picard}
}}
\source{
\url{https://hubentu.github.io/others/Rcwl_DNASeq_Align.html}

\url{https://multiqc.info}
}
\usage{
multiqc
}
\description{
The mutliQC command tool
}
\examples{
## scripts to build the pipeline
## demo("multiqc")
}
\keyword{datasets}
