% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_pipelines.R
\docType{data}
\name{Mutect2PL}
\alias{Mutect2PL}
\title{GATK4: Mutect2}
\format{A `cwlStepParam` object.
\describe{
 \item{Mutect2PL}{The best practice pipeline to Identify somatic short
variants (SNVs and Indels).}
}}
\source{
\url{https://software.broadinstitute.org/gatk/best-practices/workflow?id=11146}
}
\usage{
Mutect2PL
}
\description{
Somatic short variant discovery (SNVs + Indels)
}
\keyword{datasets}
