% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_pipelines.R
\docType{data}
\name{neusomatic}
\alias{neusomatic}
\title{neusomatic caller}
\format{A `cwlStepParam` object.
\describe{
 steps:
 \item{preprocess}{Preprocess step in call mode}
 \item{call}{Call variants}
 \item{postprocess}{Postprocess step (resolve long INDEL sequences,
report vcf)}
}}
\source{
\url{https://github.com/bioinform/neusomatic}
}
\usage{
neusomatic
}
\description{
neusomatic caller pipeline with ensemble mode.
}
\keyword{datasets}
