% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_pipelines.R
\docType{data}
\name{phaseVcf}
\alias{phaseVcf}
\title{phase VCF}
\format{A `cwlStepParam` object.
\describe{
 steps:
 \item{CombineVariants}{combine germline and somatic variant}
 \item{ReadBackedPhasing}{Phasing using GATK ReadBackedPhasing}
}}
\source{
\url{https://pvactools.readthedocs.io/en/latest/pvacseq/input_file_prep/proximal_vcf.html}
}
\usage{
phaseVcf
}
\description{
phase VCF
}
\keyword{datasets}
