% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_tools.R
\docType{data}
\name{PoN}
\alias{PoN}
\title{gatk CreateSomaticPanelOfNormals}
\format{An object of class \code{cwlParam} of length 1.}
\usage{
PoN
}
\description{
gatk CreateSomaticPanelOfNormals
}
\keyword{datasets}
