% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcwlPipelines-package.R
\docType{package}
\name{RcwlPipelines}
\alias{RcwlPipelines}
\title{RcwlPipelines}
\description{
A package for a collection of Rcwl pipelines. Currently four
pipeplines have been collected in the package.
}
