% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_tools.R
\docType{data}
\name{read_distribution}
\alias{read_distribution}
\title{read_distribution.py}
\format{An object of class \code{cwlParam} of length 1.}
\usage{
read_distribution
}
\description{
read_distribution.py
}
\keyword{datasets}
