% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_tools.R
\docType{data}
\name{ReadBackedPhasing}
\alias{ReadBackedPhasing}
\title{ReadBackedPhasing}
\format{An object of class \code{cwlParam} of length 1.}
\usage{
ReadBackedPhasing
}
\description{
ReadBackedPhasing
}
\keyword{datasets}
