% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_tools.R
\docType{data}
\name{RenameSampleInVcf}
\alias{RenameSampleInVcf}
\title{Picard RenameSampleInVcf}
\format{An object of class \code{cwlParam} of length 1.}
\usage{
RenameSampleInVcf
}
\description{
Picard RenameSampleInVcf
}
\keyword{datasets}
