% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_pipelines.R
\docType{data}
\name{RSeQC}
\alias{RSeQC}
\title{RNASeq quality control by RSeQC}
\format{A `cwlStepParam` object.
\describe{
 An RNASeq QC pipeline by RSeQC
 which contains steps:
 \item{gtfToGenePred}{GTF to GenePred format}
 \item{genePredToBed}{GenePred format to Bed format}
 \item{read_distribution}{Reads distribution over genome feature}
 \item{geneBody_coverage}{Reads coverage over gene body}
}}
\source{
\url{http://rseqc.sourceforge.net/}
}
\usage{
RSeQC
}
\description{
RNASeq pipeline by STAR and featureCounts.
}
\keyword{datasets}
