% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_tools.R
\docType{data}
\name{samtools_flagstat}
\alias{samtools_flagstat}
\title{samtools flagstat}
\format{An object of class \code{cwlParam} of length 1.}
\usage{
samtools_flagstat
}
\description{
samtools flagstat
}
\keyword{datasets}
