% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_tools.R
\docType{data}
\name{samtools_index}
\alias{samtools_index}
\title{samtools index}
\format{An object of class \code{cwlParam} of length 1.}
\usage{
samtools_index
}
\description{
samtools index
}
\keyword{datasets}
