% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_tools.R
\docType{data}
\name{samtools_mpileup}
\alias{samtools_mpileup}
\title{samtools_mpileup}
\format{An object of class \code{cwlParam} of length 1.}
\usage{
samtools_mpileup
}
\description{
samtools_mpileup
}
\keyword{datasets}
