% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchContainer.R
\name{searchContainer}
\alias{searchContainer}
\title{seawrch containers}
\usage{
searchContainer(tool, repo = "biocontainers", source = c("quay", "dockerhub"))
}
\arguments{
\item{tool}{The tool to search.}

\item{repo}{The repository to lookup.}

\item{source}{The container server to search, quay.io or dockerhub.}
}
\value{
A DataFrame contains image tag names, updated dates and
    image sizes.
}
\description{
To search container images for a tool in certain repository from
quay.io or dockerhub.
}
\examples{
searchContainer("samtools")
}
