% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_pipelines.R
\docType{data}
\name{SomaticCallers}
\alias{SomaticCallers}
\title{Combined Somatic Mutation Callers}
\format{A `cwlStepParam` object.
\describe{
 steps:
 \item{Mutect2PL}{mutect2 pipeline}
 \item{MuSE}{MuSE}
 \item{manta strelka2}{strelka2 with manta}
 \item{SomaticSniper}{SomaticSniper}
 \item{VarDict}{VarDict}
 \item{LoFreq}{LoFreq}
 \item{VarScan2}{VarScan2 pipeline}
 \item{neusomatic}{neusomatic ensemble mode}
}}
\source{
\url{https://github.com/bioinform/neusomatic}
}
\usage{
SomaticCallers
}
\description{
Combined caller pipelines with neosomatic ensemble mode.
}
\keyword{datasets}
