% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_tools.R
\docType{data}
\name{SomaticSniper}
\alias{SomaticSniper}
\title{SomaticSniper}
\format{An object of class \code{cwlParam} of length 1.}
\usage{
SomaticSniper
}
\description{
SomaticSniper
}
\keyword{datasets}
