% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_tools.R
\docType{data}
\name{tabix_index}
\alias{tabix_index}
\title{tabix index}
\format{An object of class \code{cwlParam} of length 1.}
\usage{
tabix_index

tabix_index
}
\description{
tabix index

tabix_index
}
\keyword{datasets}
