% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_tools.R
\docType{data}
\name{VarScan2_processSomatic}
\alias{VarScan2_processSomatic}
\title{VarScan2_processSomatic}
\format{An object of class \code{cwlParam} of length 1.}
\usage{
VarScan2_processSomatic
}
\description{
VarScan2_processSomatic
}
\keyword{datasets}
