% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_tools.R
\docType{data}
\name{VarScan2_somaticFilter}
\alias{VarScan2_somaticFilter}
\title{VarScan2_somaticFilter}
\format{An object of class \code{cwlParam} of length 1.}
\usage{
VarScan2_somaticFilter
}
\description{
VarScan2_somaticFilter
}
\keyword{datasets}
