% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_pipelines.R
\docType{data}
\name{VarScan2Somatic}
\alias{VarScan2Somatic}
\title{VarScan2 somatic caller}
\format{A `cwlStepParam` object.
\describe{
 VarScan2 Somatic caller pipeline,
 which contains steps:
 \item{mpileup}{mpileup by samtools}
 \item{somatic}{somatic calling by VarScan2 somatic}
 \item{processSomatic}{processSomatic by VarScan2}
 \item{somaticFilter}{Filter by VarScan2}
}}
\source{
\url{http://varscan.sourceforge.net}
}
\usage{
VarScan2Somatic
}
\description{
VarScan2 Somatic caller pipeline.
}
\keyword{datasets}
