% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_tools.R
\docType{data}
\name{vcf_expression_annotator}
\alias{vcf_expression_annotator}
\title{vcf_expression_annotator}
\format{An object of class \code{cwlParam} of length 1.}
\usage{
vcf_expression_annotator
}
\description{
vcf_expression_annotator
}
\keyword{datasets}
