% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_pipelines.R
\docType{data}
\name{vcfCoverage}
\alias{vcfCoverage}
\title{vcfCoverage}
\format{A `cwlStepParam` object.
\describe{
 steps:
 \item{decompose}{vt decompose}
 \item{readcount}{bam-readcount}
 \item{readcount_annotator_snv}{add snv counts}
 \item{readcount_annotator_indel}{add indel counts}
}}
\source{
\url{https://pvactools.readthedocs.io/en/latest/pvacseq/input_file_prep/readcounts.html}
}
\usage{
vcfCoverage
}
\description{
vcfCoverage
}
\keyword{datasets}
