% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_pipelines.R
\docType{data}
\name{vcfExpression}
\alias{vcfExpression}
\title{vcfExpression}
\format{A `cwlStepParam` object.
\describe{
 steps:
 \item{kallistoQuant}{transcript quantification using kallisto quant}
 \item{cleanExp}{R function to clean results from kallisto}
 \item{vcf_expression_annotator}{add expression to vcf}
}}
\source{
\url{https://pvactools.readthedocs.io/en/latest/pvacseq/input_file_prep/expression.html}
}
\usage{
vcfExpression
}
\description{
vcfExpression
}
\keyword{datasets}
