% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allclasses.R
\name{[,cwlHub,ANY,ANY,ANY-method}
\alias{[,cwlHub,ANY,ANY,ANY-method}
\alias{title}
\alias{Command}
\alias{Container}
\alias{Type}
\title{extract}
\usage{
\S4method{[}{cwlHub,ANY,ANY,ANY}(x, value)

title(object)

Command(object)

Container(object)

Type(object)
}
\arguments{
\item{x}{A `cwlHub` object.}

\item{value}{The "BFC" ID to extract the subset.}

\item{object}{A `cwlHub` object.}
}
\description{
extract

title

Command

Container

Type
}
\examples{
\dontrun{
tools <- cwlUpdate()
title(tools)
}
}
