% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwlInstall.R
\name{cwlInstall}
\alias{cwlInstall}
\title{cwlInstall}
\usage{
cwlInstall(rname, bfc = NULL, env = .GlobalEnv)
}
\arguments{
\item{rname}{The name or filepath of tool or pipeline to install
(`rname` or `fpath` column from the `bfc` object returned from
`cwlSearch`).}

\item{bfc}{The `BiocFileCache` object for the recipes returned from
`cwlUpdate`. The default is NULL which automatically detect the
"Rcwl" cache directory.}

\item{env}{The R enviroment to export to. The default is
`.GlobalEnv`.}
}
\description{
To source Rcwl scripts
}
\details{
Note to developers that the dependent Rcwl scripts should
    be included in the recipe with `@include` tag.
}
\examples{
\dontrun{
tls <- cwlSearch("bwa")
tls$rname
cwlInstall("tl_bwa")
cwlInstall(tls$fpath[tls$rname == "tl_bwa"])  ## equivalent
bwa
}
}
