% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwlSearch.R
\name{cwlSearch}
\alias{cwlSearch}
\title{cwlSearch}
\usage{
cwlSearch(keyword, bfc = NULL, type = NULL, ...)
}
\arguments{
\item{keyword}{A (vector of) character string as keywords to search
for tools or pipelines.  Will be used to match patterns
(case-insensitive) against `rname`, `rpath`, `fpath`, `Command`
and `Container` column in the `bfc` object.}

\item{bfc}{The `BiocFileCache` object for the recipes returned from
`cwlUpdate`. The default is NULL which automatically detect the
"Rcwl" cache directory.}

\item{type}{The `Type` to filter the results, "pipeline" or "tool".}

\item{...}{More options from the internal `bfcquery` function.}
}
\value{
A BiocFileCache tibble.
}
\description{
Function to search Rcwl tools and pipelines.
}
\examples{
\dontrun{
tls <- cwlSearch(c("bwa", "mem"))
data.frame(tls)
}
}
