% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwlUpdate.R
\name{cwlUpdate}
\alias{cwlUpdate}
\title{cwlUpdate}
\usage{
cwlUpdate(cachePath = "Rcwl", force = FALSE)
}
\arguments{
\item{cachePath}{The cache path of the BiocFileCache object to
store the Rcwl tools and pipelines recipes. Default is "Rcwl".}

\item{force}{Whether to clean existing recipes cache. Default is
FALSE.}
}
\description{
Function to sync and get the most updated Rcwl recipes from the
RcwlRecipes github repository.
}
\examples{
\dontrun{
tools <- cwlUpdate()
}
}
