% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcwlPipelines-package.R
\docType{package}
\name{RcwlPipelines}
\alias{RcwlPipelines}
\alias{RcwlPipelines-package}
\title{RcwlPipelines}
\description{
A package for a collection of Rcwl pipelines. Currently four
pipeplines have been collected in the package.
}
\details{
\describe{
\item{\link{rnaseq_Sf}, \link{multiqc}:}{RNASeq alignment,
QC and quantification pipeline}
\item{\link{alignMerge}:}{DNASeq alignment}
\item{\link{GAlign}, \link{hapCall}, \link{jdCall}:}{GATK4
pipeline for Germline calling}
\item{\link{mc3}:}{TCGA MC3 pipeline for Somatic calling}
}
}
