% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_pipelines.R
\docType{data}
\name{rnaseq_Sf}
\alias{rnaseq_Sf}
\title{RNASeq pipeline with STAR and featureCounts}
\format{A `cwlStepParam` object.
\describe{
 An RNASeq alignment and quantification pipeline built by
`Rcwl`, which contains steps:
 \item{fastqc}{The reads QC step by fastQC}
 \item{STAR}{The alignment step by STAR}
 \item{samtools_index}{Index bam file by samtools}
 \item{samtools_flagstat}{Flag stat by samtools}
 \item{featureCounts}{Gene level quantification by featureCounts}
 \item{RSeQC}{QC for RNASeq alignments by RSeQC}
}}
\source{
\url{https://hubentu.github.io/others/Rcwl_RNASeq.html}
}
\usage{
rnaseq_Sf
}
\description{
RNASeq pipeline by STAR and featureCounts.
}
\keyword{datasets}
